+++ crates/ruma-common/src/canonical_json.rs
@@ -7,14 +7,16 @@ use serde_json::Value as JsonValue;
 
 mod value;
 
+pub use self::value::{CanonicalJsonObject, CanonicalJsonValue};
 use crate::RoomVersionId;
 #[cfg(feature = "events")]
 use crate::{
-    events::room::redaction::{OriginalRoomRedactionEvent, OriginalSyncRoomRedactionEvent},
+    events::room::redaction::{
+        OriginalRoomRedactionEvent, OriginalSyncRoomRedactionEvent, RoomRedactionEvent,
+        SyncRoomRedactionEvent,
+    },
     serde::Raw,
 };
-
-pub use self::value::{CanonicalJsonObject, CanonicalJsonValue};
 
 /// The set of possible errors when serializing to canonical JSON.
 #[cfg(feature = "canonical-json")]
@@ -147,6 +149,24 @@ impl TryFrom<&Raw<OriginalSyncRoomRedactionEvent>> for RedactedBecause {
     type Error = serde_json::Error;
 
     fn try_from(value: &Raw<OriginalSyncRoomRedactionEvent>) -> Result<Self, Self::Error> {
+        value.deserialize_as().map(Self)
+    }
+}
+
+#[cfg(feature = "events")]
+impl TryFrom<&Raw<RoomRedactionEvent>> for RedactedBecause {
+    type Error = serde_json::Error;
+
+    fn try_from(value: &Raw<RoomRedactionEvent>) -> Result<Self, Self::Error> {
+        value.deserialize_as().map(Self)
+    }
+}
+
+#[cfg(feature = "events")]
+impl TryFrom<&Raw<SyncRoomRedactionEvent>> for RedactedBecause {
+    type Error = serde_json::Error;
+
+    fn try_from(value: &Raw<SyncRoomRedactionEvent>) -> Result<Self, Self::Error> {
         value.deserialize_as().map(Self)
     }
 }